Split Dust FUSE callbacks
